requirements.txt
fastapi==0.116.1
uvicorn==0.32.1
psycopg[binary]==3.2.9
redis==5.2.0
celery==5.5.3
firebase-admin==6.9.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-auth-httplib2==0.2.0
google-api-python-client==2.176.0
pydantic-settings==2.10.1
sqlalchemy==2.0.41
alembic==1.16.4
asyncpg==0.30.0
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
psycopg2-binary==2.9.10
stripe==10.12.0
cryptography==43.0.3
python-dotenv==1.0.1
